requirements.txt
Flask
itsdangerous
oauth2client
six
python-jose
cachetools
